Aggregate data based on given grouping.

aggregate aggregates data based on the specified aggregation method.

aggregate(
  population,
  variables,
  strata,
  operation = "mean",
  univariate = TRUE,
  ...
)

Arguments

population
tbl with grouping (metadata) and observation variables.

variables
character vector specifying observation variables.

strata
character vector specifying grouping variables for aggregation.

operation
optional character string specifying method for aggregation, e.g. "mean", "median", "mean+sd". A sequence can comprise only of univariate functions.

univariate
boolean specifying whether the aggregation function is univariate or multivariate.

...
optional arguments passed to aggregation operation

Value

aggregated data of the same class as population.

Examples

population <- tibble::tibble(
  Metadata_group = c(
    "control", "control", "control", "control",
    "experiment", "experiment", "experiment",
    "experiment"
  ),
  Metadata_batch = c("a", "a", "b", "b", "a", "a", "b", "b"),
  AreaShape_Area = c(10, 12, 15, 16, 8, 8, 7, 7)
)
variables <- c("AreaShape_Area")
strata <- c("Metadata_group", "Metadata_batch")
aggregate(population, variables, strata, operation = "mean")

#> # A tibble: 4 x 3
#>   Metadata_group Metadata_batch AreaShape_Area
#>   <chr>          <chr>                   <dbl>
#> 1 control        a                        11  
#> 2 control        b                        15.5
#> 3 experiment     a                         8  
#> 4 experiment     b                         7