Husk data.

husk detects unwanted variation in the sample and removes it from the population.

husk(
  population,
  variables,
  sample,
  regularization_param = 1e-06,
  remove_signal = TRUE,
  remove_outliers = TRUE
)

Arguments

population
tbl with grouping (metadata) and observation variables.

variables
character vector specifying observation variables.

sample
tbl containing sample that is used by the method to estimate husking parameters. sample has same structure as population. Typically, sample corresponds to controls in the experiment.

regularization_param
optional parameter used in husking to offset eigenvalues to avoid division by zero. Default is 1.

remove_signal
optional boolean specifying whether to husk the signal instead of only scaling it down. Default is TRUE.

remove_outliers
optional boolean specifying whether to remove outliers. Default is TRUE.

Value

transformed data of the same class as population.

Examples

population <- tibble::tibble(
  Metadata_Well = c("A01", "A02", "B01", "B02"),
  Intensity_DNA = c(8, 20, 12, 32),
  Texture_DNA = c(5, 2, 43, 13)
)
variables <- c("Intensity_DNA", "Texture_DNA")
husk(population, variables, population, 1, remove_signal = TRUE)

#> INFO [2021-04-06 00:48:28] There are 0 PCs with signal.

#> # A tibble: 4 x 3
#>   Metadata_Well     V1     V2
#>   <chr>          <dbl>  <dbl>
#> 1 A01           -0.125 -0.568
#> 2 A02            0.311 -0.203
#> 3 B01           -0.681  0.331
#> 4 B02            0.495  0.440

husk(population, variables, population, 1e-5, remove_signal = TRUE)

#> INFO [2021-04-06 00:48:28] There are 0 PCs with signal.

#> # A tibble: 4 x 3
#>   Metadata_Well     V1     V2
#>   <chr>          <dbl>  <dbl>
#> 1 A01           -0.239 -1.20 
#> 2 A02            0.594 -0.432
#> 3 B01           -1.30   0.703
#> 4 B02            0.947  0.933

husk(population, variables, population, 1, remove_signal = FALSE)

#> INFO [2021-04-06 00:48:28] There are 0 PCs with signal.

#> # A tibble: 4 x 3
#>   Metadata_Well     V1     V2
#>   <chr>          <dbl>  <dbl>
#> 1 A01           -0.125 -0.568
#> 2 A02            0.311 -0.203
#> 3 B01           -0.681  0.331
#> 4 B02            0.495  0.440

husk(population, variables, population, 1e-5, remove_signal = FALSE)

#> INFO [2021-04-06 00:48:28] There are 0 PCs with signal.

#> # A tibble: 4 x 3
#>   Metadata_Well     V1     V2
#>   <chr>          <dbl>  <dbl>
#> 1 A01           -0.239 -1.20 
#> 2 A02            0.594 -0.432
#> 3 B01           -1.30   0.703
#> 4 B02            0.947  0.933