Mark outlier rows.

mark_outlier_rows drops outlier rows.

mark_outlier_rows(
  population,
  variables,
  sample,
  strata = NULL,
  operation = "svd+iqr",
  outlier_col = "is_outlier",
  ...
)

Arguments

population
tbl with grouping (metadata) and observation variables.

variables
character vector specifying observation variables.

sample
tbl containing sample that is used by outlier removal methods to estimate parameters. sample has same structure as population. Typically, sample corresponds to controls in the experiment.

strata
optional character vector specifying grouping variables for grouping prior to outlier removal. If NULL, no stratification is performed.

operation
optional character string specifying method for outlier removal. There is currently only one option ("svd_iqr").

outlier_col
optional character string specifying the name for the column that will indicate outliers (in the output). Default "is_outlier".

...
arguments passed to outlier removal operation.

Value

population with an extra column is_outlier.

Examples

suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), 100, replace = TRUE),
  Metadata_type = sample(c("control", "trt"), 100, replace = TRUE),
  AreaShape_Area = c(rnorm(98), 20, 30),
  AreaShape_Eccentricity = rnorm(100)
)
variables <- c("AreaShape_Area", "AreaShape_Eccentricity")
strata <- c("Metadata_group")
sample <- population %>% dplyr::filter(Metadata_type == "control")
population_marked <-
  cytominer::mark_outlier_rows(
    population,
    variables,
    sample,
    strata,
    operation = "svd+iqr"
  )
population_marked %>%
  dplyr::group_by(is_outlier) %>%
  dplyr::sample_n(3)

#> # A tibble: 6 x 5
#> # Groups:   is_outlier [2]
#>   Metadata_group Metadata_type AreaShape_Area AreaShape_Eccentricity is_outlier
#>   <chr>          <chr>                  <dbl>                  <dbl> <lgl>     
#> 1 b              trt                   -0.358                 -0.620 FALSE     
#> 2 a              trt                    1.85                   0.563 FALSE     
#> 3 b              control                0.160                 -0.147 FALSE     
#> 4 b              control                1.05                  -2.08  TRUE      
#> 5 a              control               30                     -0.311 TRUE      
#> 6 a              trt                    2.39                   1.29  TRUE