Feature importance based on data entropy.

svd_entropy measures the contribution of each feature in decreasing the data entropy.

svd_entropy(variables, sample, cores = NULL)

Arguments

variables
character vector specifying observation variables.

sample
tbl containing sample used to estimate parameters.

cores
optional integer specifying number of CPU cores used for parallel computing using doParallel.

Value

data frame specifying the contribution of each feature in decreasing the data entropy. Higher values indicate more information.

Examples

sample <- tibble::tibble(
  AreaShape_MinorAxisLength = c(10, 12, 15, 16, 8, 8, 7, 7, 13, 18),
  AreaShape_MajorAxisLength = c(35, 18, 22, 16, 9, 20, 11, 15, 18, 42),
  AreaShape_Area = c(245, 151, 231, 179, 50, 112, 53, 73, 164, 529)
)
variables <- c("AreaShape_MinorAxisLength", "AreaShape_MajorAxisLength", "AreaShape_Area")
svd_entropy(variables, sample, cores = 1)

#> # A tibble: 3 x 2
#>   variable                  svd_entropy
#>   <chr>                           <dbl>
#> 1 AreaShape_MinorAxisLength     0.00174
#> 2 AreaShape_MajorAxisLength     0.00285
#> 3 AreaShape_Area               -0.0424