Measure variable importance.

variable_importance measures importance of variables based on specified methods.

variable_importance(sample, variables, operation = "replicate_correlation",
  ...)

Arguments

sample
tbl containing sample used to estimate parameters.

variables
character vector specifying observation variables.

operation
optional character string specifying method for computing variable importance. Currently, only "replicate_correlation" (default) is implemented.

...
arguments passed to variable importance operation.

Value

data frame containing variable importance measures.

Examples

set.seed(123)
x1 <- rnorm(10)
x2 <- x1 + rnorm(10) / 100
y1 <- rnorm(10)
y2 <- y1 + rnorm(10) / 10
z1 <- rnorm(10)
z2 <- z1 + rnorm(10) / 1

batch <- rep(rep(1:2, each=5), 2)

treatment <- rep(1:10, 2)

replicate_id <- rep(1:2, each=10)

sample <-
  tibble::data_frame(x = c(x1, x2), y = c(y1, y2), z = c(z1, z2),
                     Metadata_treatment = treatment,
                     Metadata_replicate_id = replicate_id,
                     Metadata_batch = batch)

head(sample)

#> # A tibble: 6 x 6
#>             x          y          z Metadata_treatment Metadata_replicate_id
#>         <dbl>      <dbl>      <dbl>              <int>                 <int>
#> 1 -0.56047565 -1.0678237 -0.6947070                  1                     1
#> 2 -0.23017749 -0.2179749 -0.2079173                  2                     1
#> 3  1.55870831 -1.0260044 -1.2653964                  3                     1
#> 4  0.07050839 -0.7288912  2.1689560                  4                     1
#> 5  0.12928774 -0.6250393  1.2079620                  5                     1
#> 6  1.71506499 -1.6866933 -1.1231086                  6                     1
#> # ... with 1 more variables: Metadata_batch <int>

# `replicate_correlation`` returns the median, min, and max 
# replicate correlation (across batches) per variable
variable_importance(sample = sample,
                    variables = c("x", "y", "z"),
                    operation = "replicate_correlation",
                    strata = c("Metadata_treatment"),
                    replicates = 2,
                    split_by = "Metadata_batch",
                    replicate_by = "Metadata_replicate_id",
                    cores = 1)

#> # A tibble: 3 x 4
#>   variable    median       min       max
#>      <chr>     <dbl>     <dbl>     <dbl>
#> 1        x 0.9999749 0.9999707 0.9999791
#> 2        y 0.9959638 0.9926341 0.9992935
#> 3        z 0.6269866 0.2899923 0.9639808